Change GetDim from returning array pointer to returning array element
Having THaArrayString return a pointer to an array of array indices
was a bad design decision. (a) it requires small heap memory
allocation for any tyype of array, (b) allows access to the raw data
and (c) offers no bounds checking and so is just a recipe for
trouble. So we replace Int_t* GetDim() with Int_t GetDim(Int_t i)
which returns the size of the i-th dimension.

Also, we get rid of the fLen member in THaArrayString for space
reasons.  We'll have many (possibly 1000s of) THaArrayStrings but not
many calls to GetLen() since fixed-length arrays are rare, plus
computing the length on the fly is immediate anyway for 1-d arrays.

Noting that multidimensional arrays are even much rarer than fixed
arrays, we store the size of 1-d array in a member variable and only
allocate a heap array if fNdim >= 2. The member variable and the
pointer share space in a union since they are mutually exclusive. All
this is presumably more efficient than using a vector<int> with few
elements.

This allows us to get rid of the fDim member in THaVar as well, which
was a problem to begin with.