Remove BdataLoc initialization dependency on TClass
TClass isn't necessarily initialized at the time BdataLoc::DoRegister
is called. So instead of a pointer to a TClass, we register the class
name and search for the corresponding class object in
THaDecData::ReadDatabase. This should be completely safe as far as
initialization order is concerned.

Also, allow commas and tabs as delimiters in the database values,
so that the following will work, for example:

D.crate = \
  syncadc1, 1, 25, 16,  \
  syncadc2, 2, 24, 48,  \
...