Use hard cut on cluster chi2/dof to THaVDCPlane::FindClusters()
Using a clustering algorithm to classify good and bad tracks doesn't
look like a good idea here after all because the number of final clusters
isn't clear, nor is their classification (all fits may be good, or none,
or only some). One could perhaps run a clustering algorithm on fit
results form a large number of events, but that would be a calibration
procedure, not an event-by-event analysis.