Store list of associated clusters with the VDC hits
This makes detection of hits shared between clusters easier.
THaVDCClusters::ClaimHits() marks all hits of the cluster as associated
with it. Hits may be associated with more than one cluster. Ambiguities
are resolved when final tracks are fit in THaVDC::ConstructTracks().
THaVDCCluster::HasSharedHits() tests if a cluster's hits are unique to it.
Clusters with the same pivot wire are considered to be identical when
testing for shared hits.