Work in progress: Rewrite of THaVDCClusterFitter fit routines
FitSimpleTrack (=Linear2ParamFit) is done

LinearClusterFitWithT0 (=Linear3ParamFit) still missing calculation
of the covariance matrix

FitNLTrack is broken and needs a serious rewrite. It never performed
the non-linear fit that was advertised; instead it uses a linear
approximation and so is just a really expensive way of doing Linear3ParamFit.
An attempt at non-linear fitting is commented out with no explanation.
Thanks guys.

All these routines still need to be tested!