dbconvert: Behavior of CopyFile::ReadDB now hopefully matches LoadDBvalue
There is one difference however: The LoadDB logic allows keys to be
undefined in certain periods when using subdirectories. dbconvert
doesn't support validity end dates so far, and so the last key valid
before a directory start time will always continue to be valid until
the first timestamp in a file in that directory. This is the intended
behavior; the behavior of the old LoadDB code is flawed and perhaps
should be fixed in the future.