Replace THaVform/THaOdata with BranchHandler classes.
Replacements for THaVform and THaOdata are THaFormula and
VariableHandler/FormulaHandler

TODO:
Some sort of Print of formulas/cuts/variables would be good

Cuts are currently implemented as formulas. In certain cases we need
vector cuts, but THaCut is always scalar. As a result, cuts don't
appear in the global cut list, gHaCuts, nor can cuts from that list be
used in THaOutput.

Caching is not fully taken advantage of: FormulaHandler caches results in
fData/fBuffer. Perhaps add GetData?  In that case, let THaVhist talk
to FormulaHandler directly, not to the underlying formula?

Eliminate repetitive code in THaOutput::InitVariables/Formulas/Cuts:
Variable, formula and cut branches are now initialized via a new
BranchHandlerFactory class hierarchy. Admittedly, at present, it
doesn't save much code and makes things harder to understand. But it
is the proper object-oriented way of doing things here, with the main
benefit that the output branch types are now almost trivially extensible.

Process variables, formulas and cuts exactly in the order defined